% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsymTreeTests.R
\name{T2}
\alias{T2}
\title{T2 test for testing equality means against tree ordered alternatives in one-way ANOVA}
\usage{
T2(sample_data, significance_level)
}
\arguments{
\item{sample_data}{list}

\item{significance_level}{numeric}
}
\value{
Critical value numeric

Test statistic value numeric

Result Character
}
\description{
T2 test for testing equality means against tree ordered alternatives in one-way ANOVA
}
\details{
Testing of H_0:mu0 = mu1 = ... = muk vs H1:mu0 <= mui (at least one strict inequality), for all i >= 1, where mui represents the population means of the i-th treatment. The input consists of two variables: sample_data and significance_level. The output consists of the critical value, the T2 test statistic value, and the result, which indicates whether to reject or not reject the null hypothesis.
}
\examples{
# Generate data
set.seed(456)
control <- rnorm(20, mean = 5)
treatment1 <- rnorm(15, mean = 7)
treatment2 <- rnorm(10, mean = 6)

sample_data <- list(control, treatment1, treatment2)

# Run LRT at 5\% significance level
T2(sample_data, 0.05)
}
\author{
Subha Halder
}
